+++ svc-lighthouse-api/src/main/java/gov/va/vro/abddataaccess/model/HealthDataAssessment.java
@@ -1,4 +1,4 @@
-package gov.va.vro.model;
+package gov.va.vro.abddataaccess.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
