+++ svc-lighthouse-api/src/main/java/gov/va/vro/abddataaccess/service/FhirClient.java
@@ -5,11 +5,11 @@ import ca.uhn.fhir.parser.IParser;
 import gov.va.vro.abddataaccess.config.properties.LighthouseProperties;
 import gov.va.vro.abddataaccess.exception.AbdException;
 import gov.va.vro.abddataaccess.model.AbdClaim;
-import gov.va.vro.model.AbdBloodPressure;
-import gov.va.vro.model.AbdCondition;
-import gov.va.vro.model.AbdEvidence;
-import gov.va.vro.model.AbdMedication;
-import gov.va.vro.model.AbdProcedure;
+import gov.va.vro.abddataaccess.model.AbdBloodPressure;
+import gov.va.vro.abddataaccess.model.AbdCondition;
+import gov.va.vro.abddataaccess.model.AbdEvidence;
+import gov.va.vro.abddataaccess.model.AbdMedication;
+import gov.va.vro.abddataaccess.model.AbdProcedure;
 import lombok.extern.slf4j.Slf4j;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.Bundle.BundleEntryComponent;
