+++ svc-lighthouse-api/src/test/java/gov/va/vro/abddataaccess/service/FhirClientTest.java
@@ -10,11 +10,11 @@ import ca.uhn.fhir.rest.api.EncodingEnum;
 import gov.va.vro.abddataaccess.config.properties.LighthouseProperties;
 import gov.va.vro.abddataaccess.exception.AbdException;
 import gov.va.vro.abddataaccess.model.AbdClaim;
-import gov.va.vro.model.AbdBloodPressure;
-import gov.va.vro.model.AbdCondition;
-import gov.va.vro.model.AbdEvidence;
-import gov.va.vro.model.AbdMedication;
-import gov.va.vro.model.AbdProcedure;
+import gov.va.vro.abddataaccess.model.AbdBloodPressure;
+import gov.va.vro.abddataaccess.model.AbdCondition;
+import gov.va.vro.abddataaccess.model.AbdEvidence;
+import gov.va.vro.abddataaccess.model.AbdMedication;
+import gov.va.vro.abddataaccess.model.AbdProcedure;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
